refactor(gabinet): extract appointments.\$appointmentId route components (closes #4453)
Split the 2927-line appointment detail route into 8 dedicated components under
src/components/gabinet/appointments/: sidebar-extra, details-tab, payments-tab,
history-tab, receipts-tab, cancel-dialog, payment-dialog, and package-usage-dialog.
Route file reduced to ~1350 lines following the same pattern as #4452 and #4455.
